refactor[#536] : msw handler 리팩토링
